combines test files from different branches
